requirements.txt
discord.py
tls-client
numpy